requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.3
colorama==0.4.6
Flask==3.0.0
Flask-WTF==1.2.2
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
mysql-connector-python==9.5.0
numpy==2.3.5
packaging==25.0
propcache==0.4.1
protobuf==4.21.12
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.0.0
requests==2.32.5
twilio==9.8.8
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.22.0
